+++ CHANGES.md
@@ -1,9 +1,20 @@
 # Changes
 
+## v0.13.0
+
+__NOTE: This release introduces breaking changes to the API!__
+
+- `hub.options` and `hub.Options` have been removed. Pass an object with the
+  event name and the `allResults` flag as the first argument to emit instead.
+- `this.options` does no longer exist. `this.allResults` is exposed directlry
+  instead.
+
 ## v0.12.0
 
 __NOTE: This release introduces breaking changes to the API!__
 
+- `hub.options` and `hub.Options` have been removed. Pass an object with the
+  event name and the `allResults` flag as the first argument to emit instead.
 - `this.args` is now an array instead of function returning a copy of the
   emitted arguments
 - `this.stopped` is now a boolean instead of a function returning a boolean